Move SplitIssueTree class into separate file (#18)
